Basic functionality complete. Add the ability to save and recall simulations. Setup for last run simulation run is saved and used when program is reopened as start state of setup panel.

+++ Simple Sim/UtilityFunctions.h
@@ -16,9 +16,9 @@
 
 +(int) signOfInt: (int) n;
 
-+(void) calcSortIndexForDoubleArray:(double *) arrayToSort 
-                     WithStartIndex:(int) startIndex 
-                        AndEndIndex:(int) endIndex
-              AndReturningSortIndex:(int *) sortIndexArray;
++(void) calcSortIndexForDoubleArray: (double *) arrayToSort
+                     WithStartIndex: (NSUInteger) startIndex
+                        AndEndIndex: (NSUInteger) endIndex
+              AndReturningSortIndex: (int *) sortIndexArray;
 
 @end
